This is the model BEFORE I change the soil adhesion to consider that children of diff ages have diff adherence and that I will calculated adherence only for the SOILED part of children's hands and det the amt ingested based on the fraction of the SOILED portion that they put in their hands.
The soil adherence papers of late have considered the fraction of hand that is covered with soil and determined their soil adherence, NOT the average soil adherence of the hand.

Tables on Leckie_2000 pg 459
Okay, so my data is that something like 0.08 mg soil/cm2 AVERAGE on the hand. Instead I need to figure out
Total mg on hand/(surface area of hand * % of hand that is contaminated with soil)
-	Children primarily soil when the front sides of their palms and their fingers touch soil (could verify with video). Therefore, we assume that the entire front of the hand is contaminated with soil and the area to which the soil adheres is 29% (UNIFORM dist 23-35%).
-
There are a couple ways to handle determining the amount of soil on hands and the amount that goes into the child’s mouth. We know that hand contact with soil does not uniformly distribute soil across the hand; different hand configurations distribute hand in different locations – these locations have soil while others have “0” mg/cm2.
There is also spatial variability to the portion of children’s hands that go into their mouths.
Okay, so let’s assume that children’s hands were contaminated on the FRONT only, and assume that the “average” mouthing contact is front partial fingers – or maybe full front palm with fingers (9-15%) to average out the front partial fingers (4.6% - 7.7%) with the full front palms (23.1-34.4%) – but remember this is the fraction of the total hand that is going into the mouth, not the fraction of the contaminated areas.
If I assume that all of the contamination is on the front, then I can figure out the % that goes in their mouth as a % of the contaminated area! For example, assume that all of the contamination occurs on the front (~29% of the total hand surface area). Then assume that only the full finger goes in the mouth (12.5% of total hand surface area). Since I know these areas overlap, the fraction of the contaminated surface area that enters the mouth is 12.5/29 = 43% of the contamination will go in their mouths. Do if one had had 10 mg of soil on it, then this 10 mg would be distributed in 29% of the total surface area, and 43% of this would be mouthed, so the child would ingest 10 * 0.43 = 4.3 mg.
What if I DON’T consider spatial distribution. Then the 10 mg would be distributed over the entire hand and based on Auyeung’s data that 14% of the hand is mouthed, the child would ingest 10 * 0.14 = 14 mg MUCH lower.